Complete the About page sufficient for Beta release and move it from under Documentation and into the primary nav bar. Resolves #47
