+++ jspm.config.js
@@ -47,12 +47,12 @@ SystemJS.config({
     "@angular/platform-browser": "npm:@angular/platform-browser@2.0.0-rc.5",
     "@angular/platform-browser-dynamic": "npm:@angular/platform-browser-dynamic@2.0.0-rc.5",
     "@angular/router": "npm:@angular/router@3.0.0-rc.2",
-    "angular2-carbonldp": "npm:angular2-carbonldp@0.3.2",
+    "angular2-carbonldp": "npm:angular2-carbonldp@0.3.3",
     "angulartics2": "npm:angulartics2@1.1.9",
     "assert": "github:jspm/nodelibs-assert@0.2.0-alpha",
     "buffer": "github:jspm/nodelibs-buffer@0.2.0-alpha",
     "carbonldp": "npm:carbonldp@0.37.0",
-    "carbonldp-panel": "npm:carbonldp-panel@0.3.1",
+    "carbonldp-panel": "local:carbonldp-panel@0.3.1",
     "child_process": "github:jspm/nodelibs-child_process@0.2.0-alpha",
     "codemirror": "npm:codemirror@5.19.0",
     "constants": "github:jspm/nodelibs-constants@0.2.0-alpha",
@@ -255,7 +255,7 @@ SystemJS.config({
       "map": {
         "isarray": "npm:isarray@1.0.0",
         "base64-js": "npm:base64-js@1.2.0",
-        "ieee754": "npm:ieee754@1.1.6"
+        "ieee754": "npm:ieee754@1.1.8"
       }
     },
     "npm:readable-stream@2.1.5": {
@@ -314,11 +314,6 @@ SystemJS.config({
         "process": "npm:process@0.11.9"
       }
     },
-    "npm:angular2-carbonldp@0.3.2": {
-      "map": {
-        "js-cookie": "npm:js-cookie@2.1.3"
-      }
-    },
     "npm:carbonldp@0.37.0": {
       "map": {
         "file-type": "npm:file-type@3.8.0"
@@ -335,6 +330,11 @@ SystemJS.config({
         "bn.js": "npm:bn.js@4.11.6",
         "brorand": "npm:brorand@1.0.6",
         "hash.js": "npm:hash.js@1.0.3"
+      }
+    },
+    "npm:angular2-carbonldp@0.3.3": {
+      "map": {
+        "js-cookie": "npm:js-cookie@2.1.3"
       }
     }
   }
