+++ package.json
@@ -103,7 +103,7 @@
       "@angular/core": "npm:@angular/core@2.0.0-rc.5",
       "@angular/forms": "npm:@angular/forms@0.3.0",
       "@angular/http": "npm:@angular/http@2.0.0-rc.5",
-      "@angular/platform-browser": "npm:@angular/platform-browser@2.0.0-rc.5",
+      "@angular/platform-browser": "npm:@angular/platform-browser@^2.0.0-rc.6",
       "@angular/platform-browser-dynamic": "npm:@angular/platform-browser-dynamic@2.0.0-rc.5",
       "@angular/router": "npm:@angular/router@3.0.0-rc.2",
       "angular2-carbonldp": "npm:angular2-carbonldp@^0.3.2",
