+++ src/app/website/about/about-carbon-ldp.view.html
@@ -1,0 +1,89 @@
+<div class="ui grid" id="content-masthead">
+	<div class="row">
+		<div class="column">
+			<img class="featured-image" src="assets/images/about-page-masthead.jpg" alt="Build it Faster. Build it Better. Build it for the Next Web."/>
+		</div>
+	</div>
+</div>
+<div class="ui main grid container">
+
+	<!--
+	<div class="row">
+		<div class="sixteen wide column">
+			<div class="ui breadcrumb">
+				<div class="active section">About</div>
+			</div>
+			<h1>About Carbon LDP</h1>
+		</div>
+	</div>
+	-->
+
+	<div class="row">
+		<div class="twelve wide column">
+
+			<article class="ui vertical segment" id="article">
+
+				<h1>About Carbon LDP</h1>
+
+				<!--
+				<div class="ui message">
+					<p class="larger">Carbon LDP is a Linked Data Platform for building web apps that manage and link data within your enterprise and across the World Wide Web.</p>
+				</div>
+				-->
+
+				<p>Carbon is a server that responds to RESTful HTTP requests for reading and writing data. Acting as an
+					authentication server, a database, and a security manager, it provides most of the server-side components
+					developers need to build robust, data-centric web apps using pure client-side code. As such, it reduces
+					the need to design, develop, deploy, and configure server-side components such as an application server,
+					a database, and custom server-side code. Carbon makes it easier and faster for <em>every-day</em> web
+					developers to build apps that process and manage data rather than just content.
+				</p>
+
+				<p>Like many popular NoSQL document-oriented databases, Carbon provides the ability to read and write
+					<em>data documents</em> using simple JavaScript objects or JSON. Carbon data, however, is stored in a
+					graph database, which is more advantageous than common document-oriented and traditional relational
+					databases in a variety of cases. Since the data is stored natively in RDF (the Resource Description Framework),
+					Carbon makes it easy to model any kind of resources without the need for rigid schemas. Need a new class,
+					instance, property, or relationship? Simply add it on-the-fly. Forget about primary and foreign key
+					relationships. Applications powered by Carbon are inherently <em>semantic</em> - capable of linking data
+					with other applications and across the World Wide Web.</p>
+
+				<p>Carbon LDP turns data graphs into documents so that working with data is like working with web pages - simple.
+					Much like web pages, clients can retrieve a Carbon document by URL. However, most web pages are semi-structured and
+					designed to be rendered in a web browser for human consumption. Carbon resources are represented by fully
+					structured data. This makes them both easily transformed into HTML for rendering in a browser and more
+					easily understood and processed by machines. This leads to a host of computational benefits and integration
+					capabilities not afforded by typical application architectures. Applications powered by Carbon are ready
+					for the next Web.</p>
+
+				<p>What's more, Carbon provides fine-grained access to the data that exists within and across multiple documents.
+					Each document acts as an end-point that can be queried using the powerful SPARQL semantic query language. To query
+					across all data in a Carbon app, you can execute a query on the root document. Or, you can limit a query by
+					executing it on any given point in a document hierarchy. SPARQL provides a full set of analytic query operations
+					as well as a graph traversal syntax for graph-centric operations.</p>
+
+				<p>Carbon LDP includes a Workbench - a graphical user interface that helps you visualize and manage data outside
+					of your custom application. The Workbench provides a convenient way to create and list applications. It also
+					features a Document Explorer for visualizing the documents, data, and links that exist in an application.
+					It features a SPARQL Client so that you can compose queries and inspect data outside of your client-side
+					application code. It also provides a utility for exporting and importing data backups. The Workbench itself
+					is open-source. Developed as a Carbon app using Angular 2, it provides example code that developers can use
+					as an example for how to build rich, single-page applications using popular front-end frameworks and Carbon LDP.</p>
+
+				<p>Carbon LDP is currently in late ALPHA-stage testing with a public BETA planned to be released soon. The Carbon
+					LDP server will be freely available for download and use, with premium support options available.
+				</p>
+
+				<!-- Trade last paragraph when BETA is available. -->
+				<!--
+				<p>We invite you to download and use the Carbon LDP public beta from this site (see: Get Started). Please provide
+				feedback and suggestions as you use the product. Premium support options available.
+				</p>
+				-->
+
+
+			</article>
+		</div>
+		<aside class="four wide column"></aside>
+	</div>
+</div>
