--- src/app/website/documentation/about-carbon-ldp/about-carbon-ldp.view.html
@@ -1,54 +1,0 @@
-<div class="ui main grid container">
-	<div class="row">
-		<div class="sixteen wide column">
-
-			<div class="ui breadcrumb">
-				<span class="section">Documentation</span>
-				<i class="right arrow icon divider"></i>
-				<div class="active section">About Carbon LDP</div>
-			</div>
-
-			<h1>About Carbon LDP</h1>
-
-		</div>
-	</div>
-
-	<div class="row">
-		<div class="twelve wide column">
-
-			<article class="ui vertical segment" id="article">
-
-				<div class="ui message">
-					<p class="larger">Carbon LDP is a Linked Data Platform for building web apps that manage and link data within your enterprise and across the World Wide Web.</p>
-				</div>
-
-				<div class="ui icon yellow message">
-					<i class="warning sign icon"></i>
-					<div class="content">
-						<div class="header">Pre-release Documentation</div>
-						<p>Although Carbon LDP is not yet released, we're publishing documentation as it is prepared so that you can
-							preview what's in store.</p>
-					</div>
-				</div>
-
-				<p>Carbon provides services for working with data through a JavaScript SDK and REST API. Data is stored in a graph
-					database, which is more advantageous than traditional relational databases in a wide variety of cases. Since the
-					data is stored natively in RDF (the Resource Description Framework), Carbon makes it easy to model any kind of
-					resources without the need for rigid schemas. Need a new class, instance, property, or relationship? Simply add
-					it on-the-fly. Forget about primary and foreign key relationships. In Carbon, everything is semantic - capable of
-					linking data within your organization and on the World Wide Web.</p>
-
-				<p>Carbon LDP turns data graphs into documents so that working with data is like working with web pages - simple. Each
-					document also acts as a container for managing members and as an access point that can be queried with SPARQL.
-					Carbon makes managing data more like managing content - easy.</p>
-
-				<p>Carbon LDP is currently in late ALPHA-stage testing with a BETA release planned to be available in July, 2016.</p>
-
-			</article>
-		</div>
-		<aside class="four wide column">
-
-		</aside>
-	</div>
-
-</div>
