+++ src/app/website/header/header.component.html
@@ -9,8 +9,6 @@
 				<div class="ui dropdown link item"><a class="item" [routerLink]="[ '/documentation' ]">Documentation</a><i class="dropdown icon"></i>
 
 					<div class="menu">
-						<a class="item" [routerLink]="[ '/documentation/about-carbon-ldp' ]">About Carbon LDP</a>
-						<div class="divider"></div>
 						<a class="home" [routerLink]="[ '/documentation/essential-concepts' ]">Essential Concepts</a>
 						<a class="item" [routerLink]="[ '/documentation/linked-data-concepts' ]">Linked Data Concepts</a>
 
@@ -27,6 +25,9 @@
 						<a class="item" [routerLink]="[ '/documentation' ]">More...</a>
 					</div>
 				</div>
+
+				<a class="item" [routerLink]="[ '/about' ]">About</a>
+
 				<!--<div *ngIf=" ! authService.isAuthenticated() " class="right menu">
 					<a *ngIf="!isActive('./Login')" class="item login">
 						Log in
@@ -80,8 +81,6 @@
 			</a>-->
 			<div class="ui dropdown link item"><a class="item" [routerLink]="[ '/documentation' ]">Documentation</a><i class="dropdown icon"></i>
 				<div class="menu">
-					<a class="item" [routerLink]="[ '/documentation/about-carbon-ldp' ]">About Carbon LDP</a>
-					<div class="divider"></div>
 					<a class="home" [routerLink]="[ '/documentation/essential-concepts' ]">Essential Concepts</a>
 					<a class="item" [routerLink]="[ '/documentation/linked-data-concepts' ]">Linked Data Concepts</a>
 
@@ -97,6 +96,7 @@
 					<a class="item" [routerLink]="[ '/documentation' ]">More...</a>
 				</div>
 			</div>
+			<a class="item" [routerLink]="[ '/about' ]">About</a>
 		</div>
 	</div>
 </div>
