+++ src/app/website/website.routing.ts
@@ -8,9 +8,10 @@ import { RegisterView } from "./register/register.view";
 import { SignupThanksView } from "./signup-thanks/signup-thanks.view";
 import { UIExamplesView } from "./ui-examples/ui-examples.view";
 import { DocumentationView } from "app/website/documentation/documentation.view";
+import { AboutCarbonLDPView } from "./about/about-carbon-ldp.view";
 // Documentation
 import { HomeView as DocumentationHomeView } from "./documentation/home/home.view";
-import { AboutCarbonLDPView } from "./documentation/about-carbon-ldp/about-carbon-ldp.view";
+
 // Documentation -> Essential Concepts
 import { EssentialConceptsView } from "./documentation/essential-concepts/essential-concepts.view";
 import { LinkedDataConceptsView } from "./documentation/essential-concepts/linked-data-concepts.view";
@@ -42,6 +43,13 @@ const websiteRoutes:Routes = [
 				component: HomeView,
 				data: {
 					title: "Home",
+				},
+			},
+			{
+				path: "about",
+				component: AboutCarbonLDPView,
+				data: {
+					title: "About",
 				},
 			},
 			{
@@ -78,13 +86,6 @@ const websiteRoutes:Routes = [
 						data: {
 							title: "Documentation"
 						}
-					},
-					{
-						path: "about-carbon-ldp",
-						component: AboutCarbonLDPView,
-						data: {
-							title: "About Carbon LDP",
-						},
 					},
 					{
 						path: "essential-concepts",
