+++ src/app/website/website.view.scss
@@ -65,4 +65,13 @@ website {
 		padding-top: 2em;
 	}
 
+	#content-masthead {
+		margin-top: 65px;
+	}
+
+	.featured-image {
+		width: 100%;
+	}
+
 }
+
